Backend/requirements.txt
firebase==3.0.1
firebase-admin==5.2.0
Pillow==9.1.0
opencv-python==4.5.5.64
scipy
numpy==1.21.2
dlib==19.23.1
mtcnn==0.1.1
matplotlib==3.5.1
torch==1.11.0
torchvision==0.12.0
google-cloud-firestore==2.4.0
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
python-jwt==3.3.2
gcloud==0.18.3
sseclient==0.0.27
pycrypto==2.6.1
requests-toolbelt==0.9.1
tensorflow
Flask
Flask-Cors